Fixed links to .csv data files
The CSV file referenced in this lab no longer exists. Upon exploring the history of the repository, I discovered that it was replaced with a condensed version in a format specific to R. 
To keep the module intact, I have changed the links to the most recently available version of the CSV file.